requirements.txt
inky[rpi]>=1.4.0
Pillow>=9.0.0
pyyaml>=6.0
requests>=2.28.0 